Split naive::time module into separate files
